feat(debugger): support complex and nested maps in variable value visualizer
Fixes #770

+++ crates/tolkc/src/source_map.rs
@@ -331,7 +331,7 @@ pub enum Declaration {
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct PrefixInfo {
     pub prefix_str: String,
-    pub prefix_len: usize,
+    pub prefix_len: i32,
 }
 
 #[derive(Clone, Serialize, Deserialize, Debug)]
